Final testing of highlighting editable cells in tables.
